# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StructuredData {
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phoneNumber: string | null;
  complaintCategory: string;
  issueDescription: string;
  resolutionProvided: string;
  complaint: boolean;
  escalationRequired: boolean;
  supportingDocumentAvailable: boolean;
  overallCaseStatus: string;
}

export interface Case {
  id: number;
  filename: string;
  fileType: string;
  status: string;
  createdAt: string;
  structuredData: StructuredData;
  customerEmail: string;
  caseSummary: string;
  sourcePreview: string;
  processingMs: number;
}

export type DocumentInputEncoding = typeof DocumentInputEncoding[keyof typeof DocumentInputEncoding];


export const DocumentInputEncoding = {
  text: 'text',
  base64: 'base64',
} as const;

export interface DocumentInput {
  /** @minLength 1 */
  filename: string;
  /** @minLength 1 */
  content: string;
  encoding: DocumentInputEncoding;
}

export interface BatchInput {
  /** @minItems 1 */
  documents: DocumentInput[];
}

export type DashboardSummaryStatusCounts = {[key: string]: number};

export type DashboardSummaryCategoryCounts = {[key: string]: number};

export interface DashboardSummary {
  totalCases: number;
  processedToday: number;
  escalations: number;
  avgProcessingMs: number;
  statusCounts: DashboardSummaryStatusCounts;
  categoryCounts: DashboardSummaryCategoryCounts;
}

export type ListCasesParams = {
status?: string;
search?: string;
};