# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cases
    description: Customer complaint case processing
  - name: dashboard
    description: Case workflow overview
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List processed complaint cases
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a processed complaint case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Case not found
  /cases/process:
    post:
      operationId: processCase
      tags: [cases]
      summary: Process one complaint document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Processed case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "400":
          description: Invalid document
  /cases/batch:
    post:
      operationId: processBatch
      tags: [cases]
      summary: Process multiple complaint documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Processed cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard metrics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StructuredData:
      type: object
      properties:
        customerName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        complaintCategory:
          type: string
        issueDescription:
          type: string
        resolutionProvided:
          type: string
        complaint:
          type: boolean
        escalationRequired:
          type: boolean
        supportingDocumentAvailable:
          type: boolean
        overallCaseStatus:
          type: string
      required:
        - customerName
        - email
        - phoneNumber
        - complaintCategory
        - issueDescription
        - resolutionProvided
        - complaint
        - escalationRequired
        - supportingDocumentAvailable
        - overallCaseStatus
    Case:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        status:
          type: string
        createdAt:
          type: string
        structuredData:
          $ref: "#/components/schemas/StructuredData"
        customerEmail:
          type: string
        caseSummary:
          type: string
        sourcePreview:
          type: string
        processingMs:
          type: integer
      required:
        - id
        - filename
        - fileType
        - status
        - createdAt
        - structuredData
        - customerEmail
        - caseSummary
        - sourcePreview
        - processingMs
    DocumentInput:
      type: object
      properties:
        filename:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        encoding:
          type: string
          enum: [text, base64]
      required:
        - filename
        - content
        - encoding
    BatchInput:
      type: object
      properties:
        documents:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DocumentInput"
      required:
        - documents
    DashboardSummary:
      type: object
      properties:
        totalCases:
          type: integer
        processedToday:
          type: integer
        escalations:
          type: integer
        avgProcessingMs:
          type: integer
        statusCounts:
          type: object
          additionalProperties:
            type: integer
        categoryCounts:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalCases
        - processedToday
        - escalations
        - avgProcessingMs
        - statusCounts
        - categoryCounts