# lib/api-zod/src/generated/types/batchInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentInput } from './documentInput';

export interface BatchInput {
  /** @minItems 1 */
  documents: DocumentInput[];
}