# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StructuredData } from './structuredData';

export interface Case {
  id: number;
  filename: string;
  fileType: string;
  status: string;
  createdAt: string;
  structuredData: StructuredData;
  customerEmail: string;
  caseSummary: string;
  sourcePreview: string;
  processingMs: number;
}