# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryCategoryCounts } from './dashboardSummaryCategoryCounts';
import type { DashboardSummaryStatusCounts } from './dashboardSummaryStatusCounts';

export interface DashboardSummary {
  totalCases: number;
  processedToday: number;
  escalations: number;
  avgProcessingMs: number;
  statusCounts: DashboardSummaryStatusCounts;
  categoryCounts: DashboardSummaryCategoryCounts;
}