# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentInputEncoding } from './documentInputEncoding';

export interface DocumentInput {
  /** @minLength 1 */
  filename: string;
  /** @minLength 1 */
  content: string;
  encoding: DocumentInputEncoding;
}