# lib/api-zod/src/generated/types/structuredData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StructuredData {
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phoneNumber: string | null;
  complaintCategory: string;
  issueDescription: string;
  resolutionProvided: string;
  complaint: boolean;
  escalationRequired: boolean;
  supportingDocumentAvailable: boolean;
  overallCaseStatus: string;
}